Revert "Raise an exception if a database connection is not established before using acts_as_authentic"
Monkeypatch Authlogic to remove exception on database setup check.

This is monkeypatch of Authlogic's acts_as_authentic that prevents generating an
exception on configuring authlogic on model which does not have appropriate
setup in database.

Usually everything works fine with this check because a model is loaded later in
development mode. However when models are loaded early because of observers or
using ActiveAdmin this check will raise exception everytime when we just call
rake db:setup.

I think that this should be fixed on Authlogic side and I am going to write
minimal test case for that.

This reverts commit ed6d8f3a5c5947a9a5aa13654870cea628565ca0.